releaseタグの取得先を個人のサーバ経由に変更
