+++ electron-src/types/GitHub.d.ts
@@ -73,16 +73,6 @@ export interface GithubIssueReaction {
 	total_count: number;
 }
 
-export interface GithubRelease {
-	id: number;
-	node_id: string;
-	tag_name: string;
-	name: string;
-	html_url: string;
-	body: string;
-	created_at: string;
-}
-
 export interface GithubPrReview {
 	id: number;
 	node_id: string;
