+++ electron-src/types/Walk8243.d.ts
@@ -1,0 +1,6 @@
+export interface Release {
+	key: string;
+	tag: string;
+	title: string;
+	url: string;
+}
